state types refactored out of single class
